requirements.dev.txt
pre-commit>=3.6,<4.0
black>=24.8.0,<25.0
flake8>=7.1.0,<8.0
isort>=5.13.2,<6.0
mypy>=1.10,<1.11
types-requests>=2.32.0.20240712
pytest-cov>=4.1,<5.0
